gh-8 remove nfs provisioner
In the previous release we thought it was a good idea to bundle the nfs
provisioner which is required for the etcd distasterRecovery with the
chart. However, as we have now realized the lifecycles are drastically
different. In fact, deleting the chart creates circular dependencies,
because it would also delete the provisioner, thus making it impossible
to delete volumes requested by other pods of the chart.