feat: Add anti-affinity labels to ensure 1 GPU-enabled pod per node
Due to the lack of support for GPU virtualization by Weaviate (see
https://github.com/semi-technologies/t2v-transformers-models/issues/39
for details), we need to ensure each pod that uses the GPU, utilizes the
full pod without sharing it.

The way to ensure this is to schedule each GPU enabled pod to its own
dedicated node using Kubernetes's anti-affinity feature.

This addresses this to make it easy to run clusters with GPU support.