Add label to weaviate service for monitoring.
Currently, the ServiceMonitor uses a set of labels which
match all the three existing services: weaviate, weaviate-headless
and weaviate-grpc. This ends up into importing three different
targets in Prometheus, multiplying all the metrics by 3.
This PR adds a new label monitoring: enabled in the weaviate
service only if serviceMonitor is enabled (and the PROMETHEUS env
variable too). Like this, only one target will be imported, reducing
the number of metrics.